+++ plugins/nf-schema/src/main/nextflow/validation/SchemaValidator.groovy
@@ -40,7 +40,7 @@ import org.yaml.snakeyaml.Yaml
 @CompileStatic
 class SchemaValidator extends PluginExtensionPoint {
 
-    static final List<String> NF_OPTIONS = [
+    final List<String> NF_OPTIONS = [
             // Options for base `nextflow` command
             'bg',
             'c',
@@ -224,8 +224,9 @@ class SchemaValidator extends PluginExtensionPoint {
         final Path schema,
         final Map options = null
     ) {
-        def SamplesheetConverter converter = new SamplesheetConverter(samplesheet, schema, config, options)
-        return converter.validateAndConvertToList()
+        def SamplesheetConverter converter = new SamplesheetConverter(config)
+        def List output = converter.validateAndConvertToList(samplesheet, schema, options)
+        return output
     }
 
     //
@@ -431,6 +432,19 @@ Please contact the pipeline maintainer(s) if you see this warning as a user.
         def Map paramsMap = Utils.paramsLoad( Path.of(Utils.getSchemaPath(session.baseDir.toString(), schemaFilename)) )
         for (group in paramsMap.keySet()) {
             def Map groupSummary = getSummaryMapFromParams(params, paramsMap.get(group) as Map)
+            config.summary.hideParams.each { hideParam ->
+                def List<String> hideParamList = hideParam.tokenize(".") as List<String>
+                def Integer indexCounter = 0
+                def Map nestedSummary = groupSummary
+                if(hideParamList.size() >= 2 ) {
+                    hideParamList[0..-2].each { it ->
+                        nestedSummary = nestedSummary?.get(it, null)
+                    }
+                }
+                if(nestedSummary != null ) {
+                    nestedSummary.remove(hideParamList[-1])
+                }
+            }
             paramsSummary.put(group, groupSummary)
         }
         paramsSummary.put('Core Nextflow options', workflowSummary)
@@ -542,7 +556,7 @@ Please contact the pipeline maintainer(s) if you see this warning as a user.
     //
     // Clean and check parameters relative to Nextflow native classes
     //
-    private static Map cleanParameters(Map params) {
+    private Map cleanParameters(Map params) {
         def Map new_params = (Map) params.getClass().newInstance(params)
         for (p in params) {
             // remove anything evaluating to false